Finally using fusedClient with LocationRequest! (App is "working"? but doesn't do anything).
